Linux: Use an external player to make it possible to interrupt the speech. Fixes #3167.
Instead of having a thread call the synchronous function festival_say_text,
we now use festival_text_to_wave() and write the voice into a temporary file.
Then, we fork an external process to play the voice (using aplay, by default).
This makes it possible to stop the process to interrupt the voice when a
new sentence has to be said.
The external process may be configured through the environment variable
QTSPEECH_CMD, default value: "aplay %f". The '%f' token is replaced by the
actual path to the WAV file.